requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
python-dotenv
asyncpg
aiohttp
gspread
google-auth